Rename actions back to screens.
