+++ lib/screens/lister.rb
@@ -1,4 +1,4 @@
-require 'actions/action'
+require 'screens/action'
 
 class Lister < Action
