+++ lib/screens/updater.rb
@@ -1,4 +1,4 @@
-require 'actions/action'
+require 'screens/action'
 
 class Updater < Action
