+++ spec/screens/lister_spec.rb
@@ -1,6 +1,6 @@
 require 'ui/ui'
 require 'db/db_file'
-require 'actions/lister'
+require 'screens/lister'
 
 describe Lister do
